Create a separate module for storing the dagger provider.
