+++ app/src/main/java/org/groundplatform/android/GroundApplicationModule.kt
@@ -25,15 +25,6 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import java.util.Locale
 import javax.inject.Singleton
-import org.groundplatform.android.data.repository.LocationOfInterestRepositoryImpl
-import org.groundplatform.android.domain.repository.LocationOfInterestRepository
-import org.groundplatform.android.persistence.local.stores.LocalLocationOfInterestStore
-import org.groundplatform.android.persistence.local.stores.LocalSurveyStore
-import org.groundplatform.android.persistence.remote.RemoteDataStore
-import org.groundplatform.android.persistence.sync.MutationSyncWorkManager
-import org.groundplatform.android.persistence.uuid.OfflineUuidGenerator
-import org.groundplatform.android.repository.UserRepository
-import org.groundplatform.android.system.auth.AuthenticationManager
 import org.groundplatform.android.ui.common.ViewModelModule
 
 @InstallIn(SingletonComponent::class)
@@ -52,26 +43,4 @@ object GroundApplicationModule {
   }
 
   @Provides fun provideLocale() = Locale.getDefault()
-
-  @Provides
-  @Singleton
-  fun provideLocationOfInterestRepository(
-    authenticationManager: AuthenticationManager,
-    localLoiStore: LocalLocationOfInterestStore,
-    localSurveyStore: LocalSurveyStore,
-    mutationSyncWorkManager: MutationSyncWorkManager,
-    remoteDataStore: RemoteDataStore,
-    userRepository: UserRepository,
-    uuidGenerator: OfflineUuidGenerator,
-  ): LocationOfInterestRepository {
-    return LocationOfInterestRepositoryImpl(
-      authenticationManager,
-      localLoiStore,
-      localSurveyStore,
-      mutationSyncWorkManager,
-      remoteDataStore,
-      userRepository,
-      uuidGenerator,
-    )
-  }
 }
